refactor: 팀 피드백 반영 - UUID 타입 및 네이밍 컨벤션 적용

+++ backend/src/main/java/org/sejongisc/backend/attendance/entity/Location.java
@@ -15,7 +15,7 @@ public class Location {
     private Double lat;
 
     @Column(name = "longitude", precision = 10, scale = 7)
-    private Long lng;
+    private Double lng;
 
     @Column(name = "radius_meters")
     private Integer radiusMeters;
